web.pyの分割してコードを見やすくした
