+++ score.json
@@ -1,14 +1,14 @@
 [
     {
+        "player_name": "happy",
+        "score": 2100
+    },
+    {
         "player_name": "tanaka zirou",
         "score": 1800
     },
     {
         "player_name": "suzuki tarou",
         "score": 1500
-    },
-    {
-        "player_name": "happy",
-        "score": 1100
     }
 ]
